implement E2E for CNB exchange rate provider

+++ Mews.BackendDeveloperTask/Mews.BackendDeveloperTask.ExchangeRates/ExchangeRate.cs
@@ -1,3 +1,3 @@
 namespace Mews.BackendDeveloperTask.ExchangeRates;
 
-public record ExchangeRate(Currency Source, Currency Target, float Rate);
+public record ExchangeRate(Currency Source, Currency Target, decimal Rate);
